changes on refactored code
