+++ TypeScript/test/golden-master-text-test.ts
@@ -1,4 +1,4 @@
-import { Item, GildedRose } from '../app/gilded-rose-check';
+import { Item, GildedRose } from '../app/gilded-rose';
 
 const items = [
   new Item("+5 Dexterity Vest", 10, 20), //
